commit cache invalidation functions
I've copy-pasta-ed the functions used for cache invalidation from WPCOM with a few small tweaks. This probably couldn't become a core patch without a lot of work since will return lies for get_lastpostmodified() that we find acceptable. (only when a post is last modified and it impacts feeds). It might be better to look into converting the wp_cache_get / set inside _get_last_post_time() to store as a transient. Another possibility is to change the queries to query off something else that would have an index. Not sure how feasible that is.